Pull keyboard event translation into core
