+++ src/core/Types.ts
@@ -1,0 +1,10 @@
+/**
+ * Copyright (c) 2018 The xterm.js authors. All rights reserved.
+ * @license MIT
+ */
+
+export interface IKeyboardResult {
+  cancel: boolean;
+  key: string;
+  scrollLines: number;
+}
